api/requirements.txt
uvicorn == 0.18.2
fastapi == 0.95.1
tensorflow == 2.8.0
python-multipart == 0.0.6
Pillow == 9.5.0
numpy == 1.24.3
protobuf == 3.19.4
